+++ Source/CodeEntry.h
@@ -117,7 +117,7 @@ private:
 
    void OnClicked(float x, float y, bool right) override;
    bool MouseMoved(float x, float y) override;
-   bool MouseScrolled(int x, int y, float scrollX, float scrollY) override;
+   bool MouseScrolled(float x, float y, float scrollX, float scrollY) override;
 
    static bool sWarnJediNotInstalled;
